+++ app/gui/view/graph-editor/src/builtin/visualization/java_script/table.js
@@ -1,7 +1,6 @@
 /** Table visualization. */
-loadScript('https://cdn.jsdelivr.net/npm/ag-grid-community/dist/ag-grid-community.min.js')
-// Use the following line instead of the above one to use the enterprise version of ag-grid.
-// loadScript('https://cdn.jsdelivr.net/npm/ag-grid-enterprise@29.1.0/dist/ag-grid-enterprise.min.js')
+
+loadScript('https://cdn.jsdelivr.net/npm/ag-grid-enterprise/dist/ag-grid-enterprise.min.js')
 
 // ============================
 // === Style Initialisation ===
@@ -153,6 +152,13 @@ class TableVisualization extends Visualization {
                 onColumnResized: e => this.lockColumnSize(e),
                 suppressFieldDotNotation: true,
             }
+
+            if (typeof AG_GRID_LICENSE_KEY !== 'undefined') {
+                agGrid.LicenseManager.setLicenseKey(AG_GRID_LICENSE_KEY)
+            } else {
+                console.warn('The AG_GRID_LICENSE_KEY is not defined.')
+            }
+
             this.agGrid = new agGrid.Grid(tabElem, this.agGridOptions)
         }