+++ app/gui2/src/components/GraphEdge.vue
@@ -50,6 +50,7 @@ const edgePath = computed(() => {
 .edge {
   stroke: tan;
 }
+
 .edge:hover {
   stroke: red;
 }